backend/requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Database
pymongo==4.6.1

# Google Gemini API
google-generativeai>=0.5.0

# Environment Management
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
